# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/webhooks/bitbucket": {
      "post": {
        "operationId": "WebhooksController_bitbucket",
        "summary": "Endpoint for Bitbucket webhooks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook received and processed successfully"
          },
          "401": {
            "description": "Invalid or missing signature"
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bitbucket-webhook": []
          }
        ]
      }
    },
    "/api/webhooks/github": {
      "post": {
        "operationId": "WebhooksController_github",
        "summary": "Endpoint for GitHub webhooks",
        "parameters": [
          {
            "name": "X-GitHub-Event",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook received and processed successfully"
          },
          "401": {
            "description": "Invalid or missing signature"
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "github-webhook": []
          }
        ]
      }
    },
    "/api/webhooks/gitlab": {
      "post": {
        "operationId": "WebhooksController_gitlab",
        "summary": "Endpoint for GitLab webhooks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook received and processed successfully"
          },
          "401": {
            "description": "Invalid or missing signature"
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "gitlab-webhook": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "achievibit-api",
    "description": "![swagger-mode](https://img.shields.io/badge/mode-devcontainer-FF5BF8)\n\nSince this swagger shares the same domain as the app, you can use the same cookie for authentication.\n\nFor the WebSocket API, please visit [achievibit-ws](/api/docs-async)",
    "version": "1.0",
    "contact": {
      "name": "thatkookooguy",
      "url": "https://github.com/thatkookooguy",
      "email": "thatkookooguy@kibibit.io"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT token for authenticated user routes"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user_token",
        "description": "JWT token for authenticated user routes"
      },
      "gitlab-webhook": {
        "type": "apiKey",
        "name": "x-gitlab-token",
        "in": "header",
        "description": "Secret token for GitLab webhooks validation (usually sent by GitLab)"
      },
      "github-webhook": {
        "type": "apiKey",
        "name": "x-hub-signature-256",
        "in": "header",
        "description": "Secret token for GitHub webhooks validation (usually sent by GitHub)"
      },
      "bitbucket-webhook": {
        "type": "apiKey",
        "name": "x-hub-signature",
        "in": "header",
        "description": "Secret token for Bitbucket webhooks validation (usually sent by Bitbucket)"
      }
    },
...
  }
}